Implemented experimental sniffer api.
