+++ esp-wifi/src/esp_now/mod.rs
@@ -19,7 +19,7 @@ use crate::{
     binary::include::*,
     compat::queue::SimpleQueue,
     hal::peripheral::{Peripheral, PeripheralRef},
-    wifi::Protocol,
+    wifi::{Protocol, RxControlInfo},
     EspWifiInitialization,
 };
 
@@ -177,57 +177,6 @@ pub struct PeerInfo {
     pub channel: Option<u8>,
     pub encrypt: bool,
     // we always use STA for now
-}
-
-#[cfg(not(any(esp32c6)))]
-#[derive(Debug, Clone, Copy)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct RxControlInfo {
-    pub rssi: i32,
-    pub rate: u32,
-    pub sig_mode: u32,
-    pub mcs: u32,
-    pub cwb: u32,
-    pub smoothing: u32,
-    pub not_sounding: u32,
-    pub aggregation: u32,
-    pub stbc: u32,
-    pub fec_coding: u32,
-    pub sgi: u32,
-    pub ampdu_cnt: u32,
-    pub channel: u32,
-    pub secondary_channel: u32,
-    pub timestamp: u32,
-    pub noise_floor: i32,
-    pub ant: u32,
-    pub sig_len: u32,
-    pub rx_state: u32,
-}
-
-#[cfg(esp32c6)]
-#[derive(Debug, Clone, Copy)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct RxControlInfo {
-    pub rssi: i32,
-    pub rate: u32,
-    pub sig_len: u32,
-    pub rx_state: u32,
-    pub dump_len: u32,
-    pub he_sigb_len: u32,
-    pub cur_single_mpdu: u32,
-    pub cur_bb_format: u32,
-    pub rx_channel_estimate_info_vld: u32,
-    pub rx_channel_estimate_len: u32,
-    pub second: u32,
-    pub channel: u32,
-    pub data_rssi: i32,
-    pub noise_floor: u32,
-    pub is_group: u32,
-    pub rxend_state: u32,
-    pub rxmatch3: u32,
-    pub rxmatch2: u32,
-    pub rxmatch1: u32,
-    pub rxmatch0: u32,
 }
 
 #[derive(Debug, Clone, Copy)]
@@ -813,52 +762,7 @@ unsafe extern "C" fn rcv_cb(
     ];
 
     let rx_cntl = (*esp_now_info).rx_ctrl;
-    #[cfg(not(any(esp32c6)))]
-    let rx_control = RxControlInfo {
-        rssi: (*rx_cntl).rssi(),
-        rate: (*rx_cntl).rate(),
-        sig_mode: (*rx_cntl).sig_mode(),
-        mcs: (*rx_cntl).mcs(),
-        cwb: (*rx_cntl).cwb(),
-        smoothing: (*rx_cntl).smoothing(),
-        not_sounding: (*rx_cntl).not_sounding(),
-        aggregation: (*rx_cntl).aggregation(),
-        stbc: (*rx_cntl).stbc(),
-        fec_coding: (*rx_cntl).fec_coding(),
-        sgi: (*rx_cntl).sgi(),
-        ampdu_cnt: (*rx_cntl).ampdu_cnt(),
-        channel: (*rx_cntl).channel(),
-        secondary_channel: (*rx_cntl).secondary_channel(),
-        timestamp: (*rx_cntl).timestamp(),
-        noise_floor: (*rx_cntl).noise_floor(),
-        ant: (*rx_cntl).ant(),
-        sig_len: (*rx_cntl).sig_len(),
-        rx_state: (*rx_cntl).rx_state(),
-    };
-
-    #[cfg(esp32c6)]
-    let rx_control = RxControlInfo {
-        rssi: (*rx_cntl).rssi(),
-        rate: (*rx_cntl).rate(),
-        sig_len: (*rx_cntl).sig_len(),
-        rx_state: (*rx_cntl).rx_state(),
-        dump_len: (*rx_cntl).dump_len(),
-        he_sigb_len: (*rx_cntl).he_sigb_len(),
-        cur_single_mpdu: (*rx_cntl).cur_single_mpdu(),
-        cur_bb_format: (*rx_cntl).cur_bb_format(),
-        rx_channel_estimate_info_vld: (*rx_cntl).rx_channel_estimate_info_vld(),
-        rx_channel_estimate_len: (*rx_cntl).rx_channel_estimate_len(),
-        second: (*rx_cntl).second(),
-        channel: (*rx_cntl).channel(),
-        data_rssi: (*rx_cntl).data_rssi(),
-        noise_floor: (*rx_cntl).noise_floor(),
-        is_group: (*rx_cntl).is_group(),
-        rxend_state: (*rx_cntl).rxend_state(),
-        rxmatch3: (*rx_cntl).rxmatch3(),
-        rxmatch2: (*rx_cntl).rxmatch2(),
-        rxmatch1: (*rx_cntl).rxmatch1(),
-        rxmatch0: (*rx_cntl).rxmatch0(),
-    };
+    let rx_control = RxControlInfo::from_raw(rx_cntl);
 
     let info = ReceiveInfo {
         src_address: src,
